Move BearerToken to common testing project
