--- test/Ocelot.AcceptanceTests/BearerToken.cs
@@ -1,15 +1,0 @@
-using Newtonsoft.Json;
-
-namespace Ocelot.AcceptanceTests;
-
-public class BearerToken
-{
-    [JsonProperty("access_token")]
-    public string AccessToken { get; set; }
-
-    [JsonProperty("expires_in")]
-    public int ExpiresIn { get; set; }
-
-    [JsonProperty("token_type")]
-    public string TokenType { get; set; }
-}
