+++ test/Ocelot.Testing/BearerToken.cs
@@ -1,0 +1,15 @@
+using Newtonsoft.Json;
+
+namespace Ocelot.Testing;
+
+public class BearerToken
+{
+    [JsonProperty("access_token")]
+    public string? AccessToken { get; set; }
+
+    [JsonProperty("expires_in")]
+    public int ExpiresIn { get; set; }
+
+    [JsonProperty("token_type")]
+    public string? TokenType { get; set; }
+}
